# api/openapi/cats-openapi-v1.0.1.yaml
openapi: 3.0.3
info:
  title: CATs Platform Core REST API
  description: |
    Universal Computer-Assisted Translation SaaS API Gateway Contracts

    **版本**: v1.0.1 (per §8.5 决议 10 错误码引用闭环前置同步)
    **升版触发**: per `doc/03-详细设计/接口设计/CATs_接口设计书_v2.0.md` §3.5 + 启动会决议 10 (T-07 9/27)
    **基线**: v1.0.0 (`cats-openapi-v1.yaml`)
    **改动**: 加 `components.schemas.ErrorBody` (3 字段, 跟 Rust impl `crates/auth-service/src/models.rs:42` 对齐)
             + `components.schemas.ErrorCode` enum (28 条, 引用 `doc/05-其他/管理/CATs_错误码表_v1.0.md` §3)
    **2026-10-07 追加（未升版，刻意）**: `/healthz` 的响应体统一为
             `components.schemas.HealthResponse`，并移除旧的顶层 `service` /
             `name` / `version`。这是**破坏性变更**，按 `CATs_接口设计书_v2.0.md`
             §7 接口版本控制应当走 minor 升版（1.0.1 → 1.1.0），但该流程会牵动
             约 20 处跨文档引用（proto 注释 / 各 crate lib.rs 头 / README /
             接口设计书 §9 关联表），故**本轮只订正内容、暂不动版本号**，
             升版作为独立工单待拍板。在升版完成之前，本文件的内容以本段为准。
    **接口设计书 v2.0+2 引用**: per `git log -1 --format='%H'` on `doc/03-详细设计/接口设计/CATs_接口设计书_v2.0.md`
  version: 1.0.1
servers:
  - url: https://api.cats.internal/v1
    description: Internal K3s Envoy Gateway Endpoint
paths:
...
  /projects:
    get:
      summary: List Workspace Projects
      tags: [Project]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Projects list
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create Project
      tags: [Project]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, source_lang, target_lang]
              properties:
                name: { type: string }
                source_lang: { type: string }
                target_lang: { type: string }
                tm_id: { type: string }
                termbase_id: { type: string }
      responses:
        '201':
          description: Project created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
...
  schemas:
    ErrorBody:
      type: object
      description: |
        统一错误响应 schema (per 错误码表 v1.0 §2.1 + Rust impl)
        3 字段 (error + message + detail), 跟 §1.3 高层嵌套信封 (4 字段) 不同
        §1.3 是 v2.0 大版本基线设计 (8/18 8aed5d0), §3.5 是 Sprint 1 落地 schema (per 错误码表 8/27)
        真实落地以本 schema 为准 (3 字段, snake_case 28 codes)
      required: [error, message]
      properties:
        error:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: |
            人类可读错误描述 (i18n 前缀, 默认 en-US, per 错误码表 §2.1)
        detail:
          type: string
          nullable: true
          description: |
            可选上下文 (不暴露 secret / PII, 内部 detail 走 detail 不走 message, per 错误码表 §2.1)

    # ===========================================================
    # ErrorCode enum (v1.0.1 新增, 28 条 snake_case 业务错误码)
    # 引用: doc/05-其他/管理/CATs_错误码表_v1.0.md §3.2-§3.7 (28 条枚举)
    # commit baseline: 2146f53 (T-01 实战深化, 错误码表 v1.0)
    # Single Source of Truth: 错误码表 v1.0 §3, 跟 Rust impl 一致
    # 接口设计书 v2.0+2 §3.5.4 反向引用
    # ===========================================================
    ErrorCode:
      type: string
      description: |
        业务错误码枚举 (per 错误码表 v1.0 §3, 28 条, snake_case, machine-readable)
        必须枚举本表全部 28 条值 (per 错误码表 §6.3 实施要求)
      enum:
        # 客户端请求错误 (§3.2, 5 条)
        - invalid_request
        - invalid_payload
        - invalid_header
        - invalid_query
        - unsupported_version
        # 鉴权错误 (§3.3, 8 条)
        - invalid_credentials
        - invalid_token
        - token_expired
        - token_revoked
        - invalid_token_type
        - missing_authorization
        - invalid_authorization_scheme
        - user_inactive
        # 资源错误 (§3.4, 6 条)
        - user_not_found
        - token_not_found
        - resource_not_found
        - username_conflict
        - email_conflict
        - token_already_revoked
        # 限流 / 配额 (§3.5, 2 条)
        - rate_limited
        - quota_exceeded
        # 服务端错误 (§3.6, 4 条)
        - server_error
        - server_misconfigured
        - dependency_unavailable
        - dependency_timeout
        # 业务规则 (§3.7, 2 条有效, user_inactive 已在 §3.3)
        - password_mismatch
        - operation_not_permitted
...
  responses:
    BadRequest:
      description: Validation Error (HTTP 400, per 错误码表 §3.2)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Unauthorized:
      description: Unauthorized (HTTP 401, per 错误码表 §3.3)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...