# crates/auth-service/src/models.rs
//! 业务模型 + API DTO
//!
//! 引用: proto/cats/v1/auth.proto + api/openapi/cats-openapi-v1.yaml

use serde::{Deserialize, Serialize};

/// 登录请求 (REST, per OpenAPI v1)
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct LoginRequest {
    pub username: String,
    pub password: String,
}

/// 登录响应 (REST)
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct LoginResponse {
    pub access_token: String,
    pub refresh_token: String,
    pub expires_in: i64,
    pub token_type: String,
    pub user_id: String,
    pub username: String,
}

/// 刷新请求
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct RefreshRequest {
    pub refresh_token: String,
}

/// 刷新响应
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct RefreshResponse {
    pub access_token: String,
    pub refresh_token: String,
    pub expires_in: i64,
    pub token_type: String,
}

/// 错误响应 (统一格式 per 实施前QA §3.4)
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct ErrorBody {
    pub error: String,
    pub message: String,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub detail: Option<String>,
}

/// GET /v1/auth/me 响应 (per OpenAPI v1 §3.2)
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct MeResponse {
    pub user_id: String,
    pub username: String,
    pub email: String,
}

/// JWT Claims
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct Claims {
    pub sub: String, // user_id (UUID)
    pub username: String,
    pub exp: i64,           // unix timestamp
    pub iat: i64,           // unix timestamp
    pub jti: String,        // JWT ID (UUID v4)
    pub token_type: String, // "access" or "refresh"
    #[serde(default)]
    pub roles: Vec<String>, // 用户角色列表 (per ULYS-149 第二层缺口: BFF principal.roles 依赖此字段)
    /// 逗号分隔的角色串，供**边缘代理**（envoy `jwt_authn`）直接注入
    /// `X-Cats-Roles` 用。
    ///
    /// 2026-10-07 新增。为什么不直接用 `roles`：`jwt_authn` 的 `claim_to_headers`
    /// 对 **string / int / double / bool 是原样拷贝**，而对 **array / object 会
    /// 序列化成 JSON 再 Base64 编码**（见 envoy config.proto 对 claim_to_headers
    /// 的说明）。`roles` 是数组，直接映射过去会变成
    /// `WyJVc2VyIiwiU3BvbnNvciJd`，下游按逗号切分必然解析不出任何角色 →
    /// 全员 401。
    ///
    /// 所以同一个语义在这里存两份：`roles`（数组，给 Rust 侧消费，如 cats-bff 的
    /// `principal.roles`）与 `roles_csv`（字符串，给边缘注入）。两者由
    /// `issue_jwt` 从同一份入参派生，不存在两份真相。
    #[serde(default)]
    pub roles_csv: String,
}

/// DB 实体: auth_db.users_credential
#[derive(Debug, Clone, sqlx::FromRow)]
pub struct UserCredential {
    pub id: uuid::Uuid,
    pub username: String,
    pub email: Option<String>,
    pub password_hash: String,
    pub is_active: bool,
    pub created_at: chrono::DateTime<chrono::Utc>,
    pub updated_at: chrono::DateTime<chrono::Utc>,
}

// =====================================================================
// T-01: 审计事件 + 登出 + 刷新撤销 增量
// 引用: doc/05-其他/管理/CATs_M1_Sprint1_任务拆解_v1.0.md §2 T-01
// =====================================================================

/// 审计事件结果
#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
#[serde(rename_all = "snake_case")]
pub enum AuditOutcome {
    Success,
    Failure,
}

impl AuditOutcome {
    pub fn as_str(&self) -> &'static str {
        match self {
            Self::Success => "success",
            Self::Failure => "failure",
        }
    }
}

/// 审计事件 (per T-01 + 安全要件 §6)
///
/// 事件类型枚举（参考错误码表 v1.0 §3.1, 后续在表里详列）:
/// - `login` / `login_failed` — 登录成功 / 失败
/// - `logout` — 登出
/// - `refresh` / `refresh_failed` / `refresh_revoked` — 刷新成功 / 失败 / 撤销
/// - `me_access` — 当前用户查询
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct AuditEvent {
    pub event_id: uuid::Uuid,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub user_id: Option<uuid::Uuid>,
    pub event_type: String,
    pub outcome: AuditOutcome,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub detail: Option<serde_json::Value>,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub source_ip: Option<String>,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub user_agent: Option<String>,
    pub occurred_at: chrono::DateTime<chrono::Utc>,
}

/// POST /v1/auth/logout 请求 (per 接口设计书 v2.0 §3.4)
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct LogoutRequest {
    /// 撤销的 refresh_token (jti 从中解析)
    pub refresh_token: String,
}

/// POST /v1/auth/logout 响应
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct LogoutResponse {
    pub revoked: bool,
    pub revoked_at: chrono::DateTime<chrono::Utc>,
}

/// audit_log 表 raw row (查询返回)
#[derive(Debug, Clone, sqlx::FromRow)]
pub struct AuditEventRow {
    pub event_id: uuid::Uuid,
    pub user_id: Option<uuid::Uuid>,
    pub event_type: String,
    pub outcome: String,
    pub detail: Option<serde_json::Value>,
    pub source_ip: Option<String>,
    pub user_agent: Option<String>,
    pub occurred_at: chrono::DateTime<chrono::Utc>,
}